+++ apps/app/src/app/(auth)/(layout)/private-beta/[[...private-beta]]/page.tsx
@@ -4,7 +4,7 @@ import dynamic from "next/dynamic";
 const title = "Private Beta";
 const description = "New users will be invited soon.";
 const PrivateBeta = dynamic(() =>
-  import("@/components/auth/private-beta").then((mod) => mod.PrivateBeta),
+  import("@/components/auth/private-beta-form").then((mod) => mod.PrivateBeta),
 );
 
 export const metadata: Metadata = {
