+++ apps/app/src/app/[slug]/(sidebar)/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 
-import { AppSidebar } from "@/components/app-sidebar";
-import { SidebarLayout } from "@/components/sidebar-layout";
+import { AppSidebar } from "@/components/sidebar";
+import { SidebarLayout } from "@/components/sidebar/sidebar-layout";
 import { HydrateClient, batchPrefetch, trpc } from "@/trpc/server";
 import { SidebarProvider } from "@coordinize/ui/components/sidebar";
