+++ apps/app/src/app/[slug]/settings/new-space/page.tsx
@@ -1,4 +1,4 @@
-import { CreateNewSpaceForm } from "@/components/settings/create-new-space";
+import { CreateNewSpaceForm } from "@/components/settings/spaces/create-new-space";
 import { Label } from "@coordinize/ui/components/label";
 
 export default function NewSpacePage() {
