+++ apps/app/src/components/auth/join-waitlist-form.tsx
