+++ apps/app/src/components/features/activity/activity-section.tsx
