+++ apps/app/src/components/onboarding/onboarding-stepper.tsx
