+++ apps/app/src/components/sidebar/team-switcher.tsx
