+++ apps/app/src/components/teams-sidebar/teams-sidebar.tsx
