--- apps/web/.env.example
@@ -1,14 +1,0 @@
-# Neon
-DATABASE_URL="postgres://postgres:postgres@db.localtest.me:5432/coordinize"
-
-# Upstash/Local Redis Instance
-UPSTASH_REDIS_REST_URL="http://localhost:8079"
-UPSTASH_REDIS_REST_TOKEN="example_token"
-
-# Better Auth
-BETTER_AUTH_SECRET=""
-NEXT_PUBLIC_BETTER_AUTH_URL="http://localhost:3000"
-
-# App URL
-NEXT_PUBLIC_APP_URL="http://localhost:3000"
-NEXT_PUBLIC_WEB_URL="http://localhost:3001"
