--- packages/auth/.env.example
@@ -1,2 +1,0 @@
-BETTER_AUTH_SECRET=""
-NEXT_PUBLIC_BETTER_AUTH_URL="http://localhost:3000"
