--- packages/database/.env.example
@@ -1,2 +1,0 @@
-# Neon DB
-DATABASE_URL="postgres://postgres:postgres@db.localtest.me:5432/coordinize"
